+++ examples/generate-svg.ts
@@ -1,5 +1,5 @@
 import { writeFile } from 'node:fs/promises'
-import { renderText } from 'sumijs'
+import { renderText } from '@blcklab/sumijs'
 
 const svg = renderText('SUMI', {
   font: 'slant',